requirements_updated.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
contourpy==1.3.2
crispy-bootstrap4==2024.10
cycler==0.12.1
Django==4.2.21
django-crispy-forms==2.4
fonttools==4.58.0
idna==3.10
kiwisolver==1.4.8
matplotlib==3.9.4
numpy==2.2.6
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
qrcode==8.2
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
urllib3==2.4.0
